requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
numpy
pandas
psutil
matplotlib
scikit-learn
google-cloud-storage
DateTime
memory-profiler
line-profiler
psutil